requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-chroma
chromadb
pypdf
streamlit
python-dotenv
tiktoken
